fix relabel potential bugs
